Split transport plugin into transport and macros
