+++ plugs/special/transport.py
@@ -1,13 +1,18 @@
+"""
+plugs > special > transport
 
-import channels
-import general
+Contains the definition for the transport plugin
+
+Authors:
+* Miguel Guthridge [hdsq@outlook.com.au]
+"""
+
 import transport
 import ui
 
 from typing import Any
 
 from common.extensionmanager import ExtensionManager
-from common import getContext
 from common.util.apifixes import UnsafeIndex
 from controlsurfaces import (
     ControlShadowEvent,
@@ -25,19 +30,9 @@ from controlsurfaces import (
     RecordButton,
     LoopButton,
     MetronomeButton,
-    UndoButton,
-    RedoButton,
-    UndoRedoButton,
-    SaveButton,
-    QuantizeButton,
-    SwitchActiveButton,
-    SwitchActivePluginButton,
-    SwitchActiveWindowButton,
-    SwitchActiveToggleButton,
 )
 from devices import DeviceShadow
 from plugs import SpecialPlugin
-
 from plugs.eventfilters import filterButtonLift
 
 
@@ -55,20 +50,12 @@ class Transport(SpecialPlugin):
         shadow.bindMatch(StopButton, self.stopButton, raise_on_failure=False)
         shadow.bindMatch(RecordButton, self.recButton, raise_on_failure=False)
         shadow.bindMatch(LoopButton, self.loopButton, raise_on_failure=False)
-        shadow.bindMatch(MetronomeButton, self.metroButton,
-                         raise_on_failure=False)
+        shadow.bindMatch(
+            MetronomeButton, self.metroButton, raise_on_failure=False
+        )
         shadow.bindMatches(
-            NavigationButton, self.navigationButtons, raise_on_failure=False)
-
-        # Macro buttons
-        shadow.bindMatch(UndoButton, self.undo, raise_on_failure=False)
-        shadow.bindMatch(RedoButton, self.redo, raise_on_failure=False)
-        shadow.bindMatch(UndoRedoButton, self.undoRedo, raise_on_failure=False)
-        shadow.bindMatch(SaveButton, self.save, raise_on_failure=False)
-        shadow.bindMatch(QuantizeButton, self.metroButton,
-                         raise_on_failure=False)
-        shadow.bindMatches(SwitchActiveButton,
-                           self.switchActive, raise_on_failure=False)
+            NavigationButton, self.navigationButtons, raise_on_failure=False
+        )
 
     @classmethod
     def create(cls, shadow: DeviceShadow) -> 'SpecialPlugin':
@@ -165,73 +152,6 @@ class Transport(SpecialPlugin):
             ui.enter()
         else:
             return False
-        return True
-
-    @filterButtonLift
-    def undo(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        general.undoUp()
-        return True
-
-    @filterButtonLift
-    def redo(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        general.undoDown()
-        return True
-
-    @filterButtonLift
-    def undoRedo(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        general.undo()
-        return True
-
-    @filterButtonLift
-    def save(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        transport.globalTransport(92, 1)
-        return True
-
-    @filterButtonLift
-    def quantize(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        channels.quickQuantize(channels.selectedChannel())
-        return True
-
-    @filterButtonLift
-    def switchActive(
-        self,
-        control: ControlShadowEvent,
-        index: UnsafeIndex,
-        *args: Any
-    ) -> bool:
-        c = control.getControl()
-        if isinstance(c, SwitchActivePluginButton):
-            getContext().active.toggleWindowsPlugins(True)
-        elif isinstance(c, SwitchActiveWindowButton):
-            getContext().active.toggleWindowsPlugins(False)
-        elif isinstance(c, SwitchActiveToggleButton):
-            # Toggle between windows and plugins
-            getContext().active.toggleWindowsPlugins()
         return True
